refactor: reorganize command structure and update documentation
- Move command discovery and registry to infrastructure directory
- Separate command implementations from supporting infrastructure
- Update README.md with comprehensive documentation of new features
- Update CLAUDE.md with detailed architecture and usage information
- Fix command discovery to search in commands directory instead of infrastructure
- Improve code organization with clear separation of concerns
- Add detailed documentation for dual configuration system
- Include comprehensive examples and testing instructions
