+++ src/core/commands/index.ts
@@ -1,7 +1,5 @@
 export { BaseCommand } from './BaseCommand';
 export { ExecCommand } from './ExecCommand';
 export { EnvCommand } from './EnvCommand';
-
-// Dynamic loading system
-export { CommandDiscovery } from './discover';
-export { CommandRegistry } from './registry';
+export { ConfigCommand } from './ConfigCommand';
+export { MainConfigCommand } from './MainConfigCommand';
