+++ src/core/infrastructure/index.ts
@@ -1,0 +1,3 @@
+export { CommandDiscovery } from './command-discovery';
+export { CommandRegistry } from './command-registry';
+export type { CommandRegistration } from './command-registry';
